restructure chart code
