+++ vizro-ai/src/vizro_ai/_vizro_ai.py
@@ -6,9 +6,9 @@ import plotly.graph_objects as go
 from langchain_openai import ChatOpenAI
 
 from vizro_ai.chains._llm_models import _get_llm_model
-from vizro_ai.components import GetCodeExplanation, GetDebugger
-from vizro_ai.task_pipeline._pipeline_manager import PipelineManager
-from vizro_ai.utils.helper import (
+from vizro_ai.vizro_ai_charts.components import GetCodeExplanation, GetDebugger
+from vizro_ai.vizro_ai_charts.task_pipeline._pipeline_manager import PipelineManager
+from vizro_ai.vizro_ai_charts.utils.helper import (
     DebugFailure,
     PlotOutputs,
     _debug_helper,
