+++ vizro-ai/src/vizro_ai/vizro_ai_charts/components/_base.py
