+++ vizro-ai/src/vizro_ai/vizro_ai_charts/components/visual_code.py
@@ -10,8 +10,8 @@ except ImportError:  # pragma: no cov
 from langchain_core.language_models.chat_models import BaseChatModel
 
 from vizro_ai.chains._chain_utils import _log_time
-from vizro_ai.components import VizroAiComponentBase
-from vizro_ai.schema_manager import SchemaManager
+from vizro_ai.vizro_ai_charts.components import VizroAiComponentBase
+from vizro_ai.vizro_ai_charts.schema_manager import SchemaManager
 
 # 1. Define schema
 openai_schema_manager = SchemaManager()
