+++ vizro-ai/src/vizro_ai/vizro_ai_charts/schema_manager/schema_manager.py
