+++ vizro-ai/src/vizro_ai/vizro_ai_charts/utils/safeguard.py
