+++ vizro-ai/tests/unit/vizro-ai/components/test_dataframe_craft.py
@@ -3,7 +3,7 @@ import re
 import pandas as pd
 import pytest
 from langchain_community.llms.fake import FakeListLLM
-from vizro_ai.components import GetDataFrameCraft
+from vizro_ai.vizro_ai_charts.components import GetDataFrameCraft
 
 
 def dataframe_code():
