+++ vizro-ai/tests/unit/vizro-ai/components/test_visual_code.py
@@ -1,6 +1,6 @@
 import pytest
 from langchain_community.llms.fake import FakeListLLM
-from vizro_ai.components import GetVisualCode
+from vizro_ai.vizro_ai_charts.components import GetVisualCode
 
 
 @pytest.fixture
